+++ lib/tone/useTone.ts
@@ -2,10 +2,10 @@ import { useCallback } from 'react'
 
 import { TRANSPORT_CONFIG } from '#lib/constants'
 import { ToneContextValues } from '#tone/context/ToneContextProvider'
-import useInternalToneContext from '#tone/hooks/useInternalToneContext'
+import useInternalToneContext from '#tone/internal/useInternalToneContext'
 import useInternalTransportStore, {
   TransportStoreGetter,
-} from '#tone/hooks/useInternalTransportStore'
+} from '#tone/internal/useInternalTransportStore'
 
 interface UseToneCallbacks {
   handlePlay: () => void
